All implementations and code for Nash & Gregg JEA 2010 paper
This repo contains the implementations for all the data structures used
in our 2010 ACM Journal of Experimental Algorithmics paper: "Comparing
Integer Data Structures for 32- and 64-bit Keys"

The python scripts that were used to generate experimental results are
also included.

At the time, the code was compiled with g++ v4.3.2 & v3.4.6.

Not all of this code is originally mine, but I don't believe any of the
licenses restrict me from including it here.

What I think is the most interesting data structure -- the Level and
Path Compressed Burst Trie (or, LPCB-Trie) is all my code. This is the
data structure that our paper more-or-less focused on describing. This
code is in lpcbtrie.h

I've learned quite a bit since I wrote this code, so I'm not all that
proud of it (there's a reasonable amount of commented-out nonsense, for
example). That said, it is reasonably generic -- the in-node data
structures, bucket data structures, as well as bucket bursting
strategies can all be varied via templated genericity at no run-time
cost. Much of the code is carefully optimised, especially for cache
performance, although without making use of any cache parameters. At
times, this has adversely affected its readability.

Have fun!